Version 3.8.0-rc6 / Fix for wrong file deletion.
git-svn-id: https://plugins.svn.wordpress.org/live-weather-station/trunk@2066529 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ includes/system/Storage.php
@@ -357,15 +357,17 @@ class Manager {
      */
     public static function purge() {
         $count = 0;
-        $time = time() - (86400 * get_option('live_weather_station_file_retention', 7));
-        foreach (self::extended_list_dir(false) as $file) {
-            if ($file['date'] < $time) {
-                try {
-                    wp_delete_file(self::$dir . $file['file']);
-                    $count += 1;
-                }
-                catch (\Exception $ex) {
-                    //
+        if ((int)get_option('live_weather_station_file_retention', 7) > 0) {
+            $time = time() - (86400 * get_option('live_weather_station_file_retention', 7));
+            foreach (self::extended_list_dir(false) as $file) {
+                if ($file['date'] < $time) {
+                    try {
+                        wp_delete_file(self::$dir . $file['file']);
+                        $count += 1;
+                    }
+                    catch (\Exception $ex) {
+                        //
+                    }
                 }
             }
         }
